third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-session-closed-event.html

<!DOCTYPE html>
<html>
    <head>
        <title>Test MediaKeySession closed event</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            async_test(function(test)
            {
                MediaKeys.create('org.w3.clearkey').then(function(mediaKeys) {
                    var initDataType = getInitDataType();
                    return mediaKeys.createSession(initDataType, getInitData(initDataType));
                }).then(function(mediaKeySession) {
                    // Wait for the session to be closed.
                    mediaKeySession.closed.then(function(result) {
                        assert_equals(result, undefined);
                        // Now that the session is closed, verify that the
                        // closed attribute immediately returns a fulfilled
                        // promise.
                        return mediaKeySession.closed;
                    }).then(function(result) {
                        assert_equals(result, undefined);
                        test.done();
                    });

                    // release() should result in the closed promise being
                    // fulfilled.
                    return mediaKeySession.release();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Test MediaKeySession closed event.');
        </script>
    </body>
</html>
